pyproject.toml
[tool.poetry]
name = "rfkvm"
version = "0.1.0"
description = ""
authors = ["Julia Patrin <papercrane@reversefold.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyAutoGUI = "^0.9.53"
pyserial = "^3.5"
pyserial-asyncio = "^0.6"
adafruit-board-toolkit = "^1.1.0"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"
ipdb = "^0.13.9"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
